Configure the COM port properly during auto-detection
The COM port auto-detection was using the current settings for the COM port instead of configuring them to match. It turns out the COM port gets reset to defaults after a reboot, so the auto-detection breaks.

Now we query the current config, tweak it to match the parameters used by the Arduino, and try setting before reading from the port to get the ack. If the ack fails for the current port, reset the config to what it was before the probe.