Use overlapped I/O to probe all of the COM ports at once
Startup took too long probing the COM ports one at a time. Now in the probe loop it should kick off an overlapped ReadFile call but not wait for the result. As soon as one COM port completes the read successfully and receives the proper cookie from the Arduino, we can break out of the loop and use that COM port from then on. COM ports that fail the probe should restore their prior mode settings in the ~port_resources() destructor.