Fix UAC and add the frame rate to the debug output
The previous fix for UAC prompts only worked with a breakpoint that paused execution before the next call to create_resources. If the timer tried to create the resources again while the UAC prompt was still up, it would fail to enumerate the displays and leave _displays empty but set the _acquiredResources flag. Now we bail out of create_resources if there are no displays, so it will try again on the next timer until the UAC prompt is dismissed.

I also added some debug-only trace output to say what the frame rate was. Previously I'd just set breakpoints in free_resources and inspect the member variable, this way I don't have to do that anymore.