Replace the SetTimer Win32 API with C++11 thread libraries
C++11 has std::thread, std::mutex, and std::condition_variable, it should be possible to make the update_timer class platform agnostic with all of those primitives.

By splitting the timer thread and the worker thread, I also cranked the frame rate up to almost 30 fps on my machine. The timer thread sets a condition_variable to trigger an update on the worker thread, then goes back to sleep and waits with a timeout for the timer period. The timer is not delayed by the time it take the worker thread to perform each update.