Replace static update_timer instance with std::shared_ptr
Rather than capturing raw this pointers, wrap the update_timer in std::shared_ptr and capture that in the thread lambdas. That also means that once the last reference to the shared_ptr goes away, we can free the entire update_timer. We just need to keep a static std::weak_ptr to it alive to reuse it as long as it's still alive.