Fix race condition in AttachToConsole/DetachFromConsole
Accessing the screen_samples static instance from the UI thread triggered a race condition with the worker thread running in update_timer. Instead of synchronously trying to free and reacquire resources from the UI thread, we should signal the timer to stop/start, and the timer's onUpdate and onStop callback should take care of the resource reacquisition from the worker thread.