Add JSON setting serialization with a default config file
I'm using the cpprestsdk JSON library to read and write the settings object to a JSON config file. Now that it can change the settings at runtime, I'm including a pre-built ZIP file with binaries and the default configuration pretty printed and commented. I updated the ReadMe.md file to explain how to use the pre-built binaries if the user doesn't want to use Visual Studio to rebuild their own version.