Add JSON configuration for OPC servers
It's not hooked up to anything yet, but I want to be able to drive an OPC server from the same display data eventually. The configuration is set up so that ranges of pixel samples from an AdaLight display can be mapped to ranges of pixels on a channel of the OPC server.

Since I was getting a lot of duplication in the JSON persistence, I refactored it into operator>> and operator<< overrides for each of the sub-structures in settings.