Rearrange the server configuration to remove redundancy
Instead of re-defining each of the displays in each range, I'm using indices into the AdaLight displays vector. That way we don't define a different set of displays per-range or apart from the AdaLight portion.

I also created an opc_buffer object that corresponds to the serial_buffer. Each channel should have a separate buffer to render the pixels in each frame, and after we finish rendering all of the ranges of pixels for a given channel we should send the complete frame all at once.