Minor fixes
I noticed that the previousColor entry was getting set to the wrong byte order. We read it as RGBA but we write it as BGRA (which is only consistent with the optimized byte order we use for the DirectX texture mapping).

Also, for the opc_buffer we need to keep track of the highest pixel index in each channel, not the total count of pixels. If we skip some pixels at the beginning of the range, we still need to fill them in as 0, 0, 0 to get to the offset of the pixels we actually are setting. So we need a contiguous range of pixels from [0, maxOffset - 1].