Implement OPC connections
I'm using winsock2 to communicate with the OPC server(s) specified in the config file. I'm rendering a pixel buffer for each mapped channel and sending that to the OPC server at the same time as I send the serial data to the AdaLight.

This should work with either an AdaLight device, an OPC server, or both simultaneously.