Add alpha channel support with bob_buffer
We're working on an extension to OPC which will composite multiple OPC client frames accross all channels and send the layered output to the OPC server. We haven't finalized the message format, but I can encode that in the per-server settings block and start switching to the new message format dynamically for experimentation.