Implement a Gaussian blur for the OPC pixels
Since an OPC server will often map more pixels to each range than there are samples, we end up with a lot of aliasing in the OPC output. You get stairstep effects where blocks of pixels are all the same color.

It looks much better if we can smooth that out so the boundaries are blurred. I implemented a Gaussian blur which is calculated once when the configuration is loaded and then applied to each channel as it's rendered before getting sent to the OPC server.